Move ItEmbeddedMarshallingTest to `compute` package
