+++ modules/platforms/cpp/ignite/odbc/common_types.cpp
@@ -312,7 +312,7 @@ sql_state error_code_to_sql_state(error::code code) {
         case error::code::NODE_START:
             return sql_state::SHY000_GENERAL_ERROR;
 
-        // Marshalling group. Group code: асV 22
+        // Marshalling group. Group code: 22
         case error::code::COMMON:
         case error::code::UNSUPPORTED_OBJECT_TYPE:
         case error::code::MARSHALLING_TYPE_MISMATCH:
